separate admin.py from models.py, avoid migration issues with view definition

+++ annotate/migrations/0002_add_join_sql_view.py
@@ -7,6 +7,9 @@ from django.db import models
 class Migration(SchemaMigration):
 
     def forwards(self, orm):
+        if db.backend_name != 'postgres':
+            print "Warning: Non-postgres database detected; convenience view will not be created."
+            return
         db.execute('''
 CREATE OR REPLACE VIEW annotate_annotation_joined AS 
 SELECT ct.app_label, ct.model, a.object_id, at.id AS type_id, at.name AS type_name, a.value
@@ -15,6 +18,8 @@ JOIN annotate_annotationtype at ON at.id = a.type_id
 JOIN django_content_type ct ON a.content_type_id = ct.id;''')
 
     def backwards(self, orm):
+        if db.backend_name != 'postgres':
+            return
         db.execute("DROP VIEW annotate_annotation_joined;");
